Extracted JSONRPCErrorResponse to core.ts.
