+++ .betterer.results
@@ -12,7 +12,7 @@ exports[`TypeScript Strict Mode`] = {
     "src/server/express/rest_handler.ts:1073090866": [
       [207, 50, 17, "tsc: Argument of type \'unknown\' is not assignable to parameter of type \'Message | Task | TaskArtifactUpdateEvent | TaskStatusUpdateEvent\'.", "3749434707"]
     ],
-    "src/server/transports/jsonrpc/jsonrpc_transport_handler.ts:2901163198": [
+    "src/server/transports/jsonrpc/jsonrpc_transport_handler.ts:3500259735": [
       [88, 12, 10, "tsc: Variable \'rpcRequest\' is used before being assigned.", "3927050741"]
     ],
     "src/types/converters/to_proto.ts:1276072555": [
